refactoring: 09_TbResourceController: fix bug test save resource
