+++ application/src/main/java/org/thingsboard/server/service/resource/TbResourceService.java
@@ -24,11 +24,10 @@ import org.thingsboard.server.common.data.id.TenantId;
 import org.thingsboard.server.common.data.lwm2m.LwM2mObject;
 import org.thingsboard.server.common.data.page.PageData;
 import org.thingsboard.server.common.data.page.PageLink;
-import org.thingsboard.server.service.entitiy.SimpleTbEntityService;
 
 import java.util.List;
 
-public interface TbResourceService extends SimpleTbEntityService<TbResource> {
+public interface TbResourceService {
 
     TbResource saveResource(TbResource resource) throws ThingsboardException;
